Move `ipx*.cpp|h` & `ne2000.cpp|h` to `src/hardware/network`
